+++ db/migrate/20131022135331_add_parent_id_to_products.rb
@@ -1,0 +1,5 @@
+class AddParentIdToProducts < ActiveRecord::Migration
+  def change
+    add_column :shoppe_products, :parent_id, :integer, :after => :id
+  end
+end