+++ db/migrate/20131022145653_cost_price_should_be_default_to_zero.rb
@@ -1,0 +1,9 @@
+class CostPriceShouldBeDefaultToZero < ActiveRecord::Migration
+  def up
+    change_column_default :shoppe_products, :cost_price, 0.0
+  end
+  
+  def down
+    change_column_default :shoppe_products, :cost_price, nil
+  end
+end